+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
 - Ada redemption disclaimer
 - “Terms of use” page in settings section
 - Change wallet password dialog UX improvements
+- Wallet rename
 - Multiple input and output addresses in transaction details
 - Ada redemption acceptance tests
 - Show BTC and ETC currencies as coming soon in create wallet dialog